Changing the PackIcon to no longer derive from a generic class. This was causing a XamlParseException when used in element syntax (attribute syntax worked) on a menu item in the TextBox style, when the TextBox was created on a Window in Powershell. Added additional null check to TextFieldAssist
